zero123/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.12.1
torchvision==0.13.1
albumentations==0.4.3
opencv-python==4.5.5.62
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.22.2
kornia==0.6
webdataset==0.2.5
torchmetrics==0.6.0
fire==0.4.0
gradio==3.21.0
diffusers==0.12.1
datasets[vision]==2.4.0
carvekit-colab==4.1.0
rich>=13.3.2
lovely-numpy>=0.2.8
lovely-tensors>=0.1.14
plotly==5.13.1
